+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-08-29 13:42:08
-// Commit: fix(forms): prevent double tap duplicate action in forms
+// Updated at: 2026-08-31 11:23:50
+// Commit: feat(forms): add bi-directional event bus handler for forms
 
 class formsComponentState {
-  final String commitRef = '4022';
-  final String lastMessage = 'fix(forms): prevent double tap duplicate action in forms';
-  final DateTime timestamp = DateTime.parse('2026-08-29T13:42:08+05:00');
+  final String commitRef = '4065';
+  final String lastMessage = 'feat(forms): add bi-directional event bus handler for forms';
+  final DateTime timestamp = DateTime.parse('2026-08-31T11:23:50+05:00');
 }